Poprawiono kolejne bugi

+++ app/src/main/java/com/example/energii/koszt/ui/rooms/manager/DeviceManager.java
@@ -14,7 +14,7 @@ public class DeviceManager extends SQLLiteDBHelper {
         super(context);
     }
 
-    public void addDevice(String roomName, String deviceName, long powerValue, int hour, int minutes, int deviceNumber, int colorId) throws SQLEnergyCostException.EmptyField, SQLEnergyCostException.DuplicationDevice, SQLEnergyCostException.WrongTime {
+    public void addDevice(String roomName, String deviceName, double powerValue, int hour, int minutes, int deviceNumber, int colorId) throws SQLEnergyCostException.EmptyField, SQLEnergyCostException.DuplicationDevice, SQLEnergyCostException.WrongTime {
         if (roomName.isEmpty() || deviceName.isEmpty() || powerValue == 0 || deviceNumber == 0) {
             throw new SQLEnergyCostException.EmptyField(context);
         }else if(hour == 0 && minutes == 0) {
@@ -98,7 +98,7 @@ public class DeviceManager extends SQLLiteDBHelper {
         return dbhRead.rawQuery(query, new String[]{deviceName});
     }
 
-    public void updateDevice(int deviceId, String roomName, String newDeviceName, long powerValue, int deviceNumber, int hour, int minutes, int colorId) throws SQLEnergyCostException.EmptyField, SQLEnergyCostException.DuplicationDevice, SQLEnergyCostException.WrongTime {
+    public void updateDevice(int deviceId, String roomName, String newDeviceName, Double powerValue, int deviceNumber, int hour, int minutes, int colorId) throws SQLEnergyCostException.EmptyField, SQLEnergyCostException.DuplicationDevice, SQLEnergyCostException.WrongTime {
         if (roomName.isEmpty() || newDeviceName.isEmpty() || powerValue == 0 || deviceNumber == 0) {
             throw new SQLEnergyCostException.EmptyField(context);
         }else if(hour == 0 && minutes == 0) {
