+++ app/src/main/java/com/example/energii/koszt/ui/rooms/manager/RoomEditManagerDeviceDetailsListAdapter.java
@@ -46,8 +46,8 @@ public class RoomEditManagerDeviceDetailsListAdapter extends RecyclerView.Adapte
             }
         }else if(cursor.getCount() == 1){
             cursor.moveToFirst();
-            deviceKwh.add( String.format("%."+ RoomEditManager.numberAfterDot +"f",((float)cursor.getInt(1) / 1000)) + " kWh");
-            deviceCost.add(String.format("%."+ RoomEditManager.numberAfterDot +"f", cursor.getFloat(2)).replace(",","."));
+            deviceKwh.add( String.format("%."+ RoomEditManager.numberAfterDot +"f",(cursor.getDouble(1) / 1000)) + " kWh");
+            deviceCost.add(String.format("%."+ RoomEditManager.numberAfterDot +"f", cursor.getDouble(2)).replace(",","."));
             deviceName.add(cursor.getString(0));
         }
